Cleanup useless files and codes

+++ include/utils/mecs-sparse-set.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 typedef struct {
-    uint32_t* sparse; // Sparse array
+    size_t* sparse; // Sparse array
     void* dense;      // Dense array
     size_t size;      // Size of the type in dense array
     size_t capacity;  // Length of sparse array
@@ -13,6 +13,6 @@ typedef struct {
 
 void mecs_init_sparse_set(SparseSet* sparse_set, size_t size, size_t capacity, size_t length);
 void mecs_free_sparse_set(SparseSet* sparse_set);
-void* mecs_sparse_set_at(const SparseSet* sparse_set, uint32_t index);
+void* mecs_sparse_set_at(const SparseSet* sparse_set, size_t index);
 
 #endif
